+++ crates/flipperzero/src/furi/message_queue.rs
@@ -1,11 +1,11 @@
 use core::ffi::c_void;
 use core::ptr::NonNull;
-use core::time::Duration;
 
 use flipperzero_sys as sys;
-use flipperzero_sys::furi::{duration_to_ticks, Status};
+use flipperzero_sys::furi::Status;
 
 use crate::furi;
+use crate::furi::time::Duration;
 
 /// MessageQueue provides a safe wrapper around the furi message queue primitive.
 pub struct MessageQueue<M: Sized> {
@@ -30,38 +30,36 @@ impl<M: Sized> MessageQueue<M> {
     // Attempts to add the message to the end of the queue, waiting up to timeout ticks.
     pub fn put(&self, msg: M, timeout: Duration) -> furi::Result<()> {
         let mut msg = core::mem::ManuallyDrop::new(msg);
-        let timeout_ticks = duration_to_ticks(timeout);
 
         let status: Status = unsafe {
             sys::furi_message_queue_put(
                 self.hnd.as_ptr(),
                 &mut msg as *mut _ as *const c_void,
-                timeout_ticks,
+                timeout.as_ticks(),
             )
             .into()
         };
 
-        status.err_or(())
+        let _ = status.into_result()?;
+
+        Ok(())
     }
 
     // Attempts to read a message from the front of the queue within timeout ticks.
     pub fn get(&self, timeout: Duration) -> furi::Result<M> {
-        let timeout_ticks = duration_to_ticks(timeout);
         let mut out = core::mem::MaybeUninit::<M>::uninit();
         let status: Status = unsafe {
             sys::furi_message_queue_get(
                 self.hnd.as_ptr(),
                 out.as_mut_ptr() as *mut c_void,
-                timeout_ticks,
+                timeout.as_ticks(),
             )
             .into()
         };
 
-        if status.is_ok() {
-            Ok(unsafe { out.assume_init() })
-        } else {
-            Err(status)
-        }
+        let _ = status.into_result()?;
+
+        Ok(unsafe { out.assume_init() })
     }
 
     /// Returns the capacity of the queue.
@@ -102,9 +100,7 @@ impl<M: Sized> Drop for MessageQueue<M> {
 
 #[flipperzero_test::tests]
 mod tests {
-    use core::time::Duration;
-
-    use flipperzero_sys::furi::Status;
+    use super::*;
 
     use super::MessageQueue;
 
@@ -131,7 +127,7 @@ mod tests {
         // Attempting to add another message should time out.
         assert_eq!(
             queue.put(7, Duration::from_millis(1)),
-            Err(Status::ERR_TIMEOUT),
+            Err(furi::Error::TimedOut),
         );
 
         // Removing a message from the queue frees up capacity.