+++ crates/flipperzero/src/furi/mod.rs
@@ -1,6 +1,7 @@
 //! Furi API.
 
 pub mod io;
+pub mod kernel;
 pub mod log;
 pub mod message_queue;
 pub mod rng;
@@ -15,4 +16,4 @@ use flipperzero_sys as sys;
 /// Furi Result type.
 pub type Result<T> = core::result::Result<T, Error>;
 /// Furi Error type.
-pub type Error = sys::furi::Status;
+pub type Error = sys::furi::Error;